cache rag chain for reuse
